Refactored to add the merge sort capability to MERGE. Removed SMERGE
